[tests] fix DeviceTests on Windows, maybe other platforms? (#16151)
Since ac16ce63 was merged, I found several tests related to `Window`
fail on Windows with:

    InvalidOperationException: MauiContext is null.

In this case, `AlertManager.Unsubscribe()` is called after
`Window.Handler` is null:

    internal IMauiContext MauiContext =>
        Handler?.MauiContext ?? throw new InvalidOperationException("MauiContext is null.");

I updated `AlertManager.Unsubscribe()` on each platform to check
for this case and `return` instead of throwing an exception.

I am not sure if regular apps could ever hit this issue or just the
DeviceTests. .NET MAUI project templates seemed OK, so I think this
might be due to how `CreateHandlerAndAddToWindow()` works in tests?